backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
supabase==2.4.3
apscheduler==3.10.4
python-multipart==0.0.9
